+++ anikanov.vladislav/S5/mainExtansion.hpp
@@ -6,7 +6,7 @@
 #include <binarySearchTree.hpp>
 
 namespace anikanov {
-  bool loadTreeFromFile(std::string &fileName, BinarySearchTree< int, std::string, std::less<> > &tree);
+  bool loadTreeFromFile(std::string &fileName, BinarySearchTree< long long, std::string, std::less<> > &tree);
   void printInvalidUsage(std::ostream &os, char *argv[]);
   void printInvalidMode(std::ostream &os);
   void printEmpty(std::ostream &os);
